+++ src/redistribution/redist.cxx
@@ -109,22 +109,22 @@ namespace CTF_int {
   }
 
 
-  int ** compute_bucket_offsets(distribution const & old_dist,
-                                distribution const & new_dist,
-                                int64_t const *      len,
-                                int64_t const *      old_phys_edge_len,
-                                int const *          old_virt_lda,
-                                int const *          old_offsets,
-                                int * const *        old_permutation,
-                                int64_t const *      new_phys_edge_len,
-                                int const *          new_virt_lda,
-                                int                  forward,
-                                int                  old_virt_np,
-                                int                  new_virt_np,
-                                int const *          old_virt_edge_len){
+  int64_t ** compute_bucket_offsets(distribution const & old_dist,
+                                    distribution const & new_dist,
+                                    int64_t const *      len,
+                                    int64_t const *      old_phys_edge_len,
+                                    int const *          old_virt_lda,
+                                    int64_t const *      old_offsets,
+                                    int * const *        old_permutation,
+                                    int64_t const *      new_phys_edge_len,
+                                    int const *          new_virt_lda,
+                                    int                  forward,
+                                    int                  old_virt_np,
+                                    int                  new_virt_np,
+                                    int64_t const *      old_virt_edge_len){
     TAU_FSTART(compute_bucket_offsets);
 
-    int64_t **bucket_offset; alloc_ptr(sizeof(int*)*old_dist.order, (void**)&bucket_offset);
+    int64_t **bucket_offset; alloc_ptr(sizeof(int64_t*)*old_dist.order, (void**)&bucket_offset);
 
     for (int dim = 0;dim < old_dist.order;dim++){
       alloc_ptr(sizeof(int64_t)*old_phys_edge_len[dim], (void**)&bucket_offset[dim]);
@@ -173,7 +173,7 @@ namespace CTF_int {
                        distribution const & new_dist,
                        int                  new_nvirt,
                        int                  np,
-                       int const *          old_virt_edge_len,
+                       int64_t const *      old_virt_edge_len,
                        int const *          new_virt_lda,
                        int64_t *            send_counts,
                        int64_t *            recv_counts,
@@ -210,11 +210,11 @@ namespace CTF_int {
         int i_st, vc, dim;
         int64_t *  virt_counts;
         int * old_virt_idx, * virt_rank;
-        int * idx;
+        int64_t * idx;
         int64_t idx_offset;
         int64_t *  idx_offs;
         int * spad;
-        int last_len = old_dist.pad_edge_len[old_dist.order-1]/old_dist.phase[old_dist.order-1]+1;
+        int64_t last_len = old_dist.pad_edge_len[old_dist.order-1]/old_dist.phase[old_dist.order-1]+1;
         int ntd, tid;
         ntd = omp_get_num_threads();
         tid = omp_get_thread_num();
