+++ src/tensor/algstrct.cxx
@@ -872,7 +872,7 @@ namespace CTF_int {
     sr->sort(n, ptr);
   }
 
-  void ConstPairIterator::permute(int64_t n, int order, int const * old_lens, int64_t const * new_lda, PairIterator wA){
+  void ConstPairIterator::permute(int64_t n, int order, int64_t const * old_lens, int64_t const * new_lda, PairIterator wA){
     ConstPairIterator rA = * this;
 #ifdef USE_OMP
     #pragma omp parallel for
@@ -892,7 +892,7 @@ namespace CTF_int {
 
   }
 
-  void ConstPairIterator::pin(int64_t n, int order, int const * lens, int const * divisor, PairIterator pi_new){
+  void ConstPairIterator::pin(int64_t n, int order, int64_t const * lens, int const * divisor, PairIterator pi_new){
     TAU_FSTART(pin);
     ConstPairIterator pi = *this;
     int * div_lens;
@@ -926,7 +926,7 @@ namespace CTF_int {
 
   }
 
-  void depin(algstrct const * sr, int order, int const * lens, int const * divisor, int nvirt, int const * virt_dim, int const * phys_rank, char * X, int64_t & new_nnz_B, int64_t * nnz_blk, char *& new_B, bool check_padding){
+  void depin(algstrct const * sr, int order, int64_t const * lens, int const * divisor, int nvirt, int const * virt_dim, int const * phys_rank, char * X, int64_t & new_nnz_B, int64_t * nnz_blk, char *& new_B, bool check_padding){
 
     TAU_FSTART(depin);
